% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.test.R
\name{auc.test}
\alias{auc.test}
\title{AUC based independence test}
\usage{
auc.test(testy, yhat)
}
\arguments{
\item{testy}{A binary (0 or 1) vector, which will be predicted.}

\item{yhat}{A vector, prediction probabilities of testy equals to 1.}
}
\value{
result   A list of average AUC estimator and pvalue of AUC based significance test.
}
\description{
AUC based independence test.
We evaluate the significance of the AUC based statistic (Null hypothesis: AUC = 0.5, that is to say, X have no prediction ability for Y).
}
\references{
Yi Li, Xiaoyu Liu, et al. knnAUC : k-nearest neighbors AUC test.BMC Bioinformatics.2018.

Mason, S. J. and Graham, N. E. (2002), Areas beneath the relative operating characteristics (ROC) and relative operating levels (ROL) curves: Statistical significance and interpretation. Q.J.R. Meteorol. Soc. doi:10.1256/003590002320603584
}
\author{
Yi Li, \email{liyistat@gmail.com}
}
