% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knn}
\alias{knn}
\title{K-nearest neighbors classifier}
\usage{
knn(train, test, k)
}
\arguments{
\item{train}{A training dataset containing both a continuous variable (X) and a binary (0 or 1) variable (Y).}

\item{test}{A testing dataset containing both a continuous variable (X) and a binary (0 or 1) variable (Y).}

\item{k}{A positive integer, we'll automatically find the best parameter for knn between 1 and number k.}
}
\value{
yhat A vector, prediction probabilities of testy equals to 1.
}
\description{
K-nearest neighbors classifier.
We'll automatically find the best parameter k for knn from RWeka.
}
\references{
Yi Li, Xiaoyu Liu, et al. knnAUC : k-nearest neighbors AUC test. BMC Bioinformatics. 2018.
}
\author{
Yi Li, \email{liyistat@gmail.com}
}
